memory: Moved kernel thread stack management to thread module
Previously, there was a chunk of code for managing the stacks for
kernel threads, which lived in the memory module. This made sense,
as it was all about memory management, but it was only ever used
by the thread module. This commit moves that code into the thread
module in a new stacks submodule.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>