syscall: Pass in the segment data rather than pulling directly
Previously, the syscall setup would access the CPU-local data
directly to get the current segment selectors. This works fine,
but it would be preferable if the syscall functionality didn't
depend on cpu_local.

This commit fixes that by getting the kernel to pass the segment
data to the syscalls initialiser.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>