drivers: Provide the kernel's PML4 when installing PCI drivers
PCI drivers normally need to be able to interact with the page
tables to set up MMIO. This is typically a one-off during the
installation process, rather than being needed long-term during
the driver's operation. Rather than drivers accessing the PML4
directly, it's much simpler if the installer pass a reference.

This commit does just that. There are a few remaining calls to
kernel_pml4 within the drivers, but those can be removed later.

It would be nicer to pass a trait, rather than mandating a
specific concrete type, but this is much simpler. We can always
change it later if needs be.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>