virtmem: Use the current page tables in virt_to_phys_addrs
Previously, the caller would pass in a set of page tables; typically
the kernel's initial PML4. Instead, we now read the currently active
page tables and use those. It might seem unsafe, but it's actually
ok because:

- We only read the page tables, not write.
- We block interrupts to prevent thread-safety issues.
- We return None if any of the address space provided is unmapped.

This could only cause issues (a page fault) if there's a bug in
OffsetPageTable, which feels like a sensible gamble.

To avoid allocating while we block interrupts, we pre-allocate a
vector of the maximum returned size and pass it in.

At the same time, I've also fixed the unit tests in virtmem. Notably,
they required a working allocator. We can't use virtmem's normal one,
as that requires kernel-level access to memory. Instead, we patch out
the crate allocator and defer to std instead.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>