network: Implement smoltcp::phy::Device to allow other drivers
This is a fairly big change. Since the network stack was first
written, we've struggled with the fact that smoltcp::iface::Interface
is a generic type, over the device type. This means it's not possible
to have a list of interfaces unless they all have the same device
type. To get things started, I solved this by requiring all devices
to be virtio::network::Device. This worked, but it would have
prevented us from adding support for other device types.

This moves the device management into the network module. We do
this by defining a simpler network::Device trait, which isn't
generic. We then use a wrapper type to implement smoltcp::phy::Device,
deferring to a Box<dyn network::Device>. This means that the list
of interfaces all act on the same device type (network::WrappedDevice),
while still allowing diverse network device drivers. As an added
bonus, this also makes it a bit simpler to implement new network
device drivers, as the network::Device trait is fairly straightforward,
since it can work on the assumption that it runs in the kernel.

This isn't necessarily a perfect design, but it's much better
than what we had before and should be easier to change over time.
This also means the network module no longer depends on the VirtIO
driver.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>