cpu-local: Merge into cpu crate, extract some CPU-local data
This is quite a large change.

CPU-local data

There are two main changes here. Firstly, the CPU-local data has
been moved from the multitasking::cpu_local module to the cpu
crate. Secondly, we make much greater use of the unique CPU id,
storing some per-CPU data locally in the modules that use it,
rather than centralising everything in the cpu_local/cpu module.

In particular, the current thread and the idle thread now live
in the thread module, in separate vectors (indexed by CPU id),
as does the segment data, in the segmentation module. This means
the cpu crate doesn't need to depend directly on those modules,
or on any other modules that wish to add per-CPU data later.

The only data remaining in the central per-CPU memory region
are the unique CPU id, the saved syscall stack pointer, and the
user stack pointer. The CPU id has to stay there as there's
nowhere else to put it and the two stack pointers have to stay
because we reference them without a kernel stack from assembly,
so there's no other way to access them.

The three remaining values in the central region are fixed-size
integers and not pointers, which makes accessing their values
more efficient, simpler, and safer than the previous design.

Segmentation

With the segmentation crate now owning the per-CPU segment
data, we can make its initialisation simpler and more safe. We
now just have two initialisation functions; bootstrap to prepare
the global segment data we use until we have per-CPU data set up,
then per_cpu_init to migrate to the per-CPU segmentation data.

We still perform the same checks as before that initialisation
only happens once per CPU (and once globally), but this design
allows us to hide the init/activate/swap functions, making it
even harder to get wrong.

Number of CPU cores

The main gap this commit leaves is that we don't yet have the
functionality to detect the maximum number of CPU cores the
machine supports. For now, we just hard-code it to 1, as we
don't have support for initialising other CPUs anyway. However,
the mechanisms are in place so it should be fairly easy to fix
later.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>