scheduler: Reduced race conditions between resume and suspend/sleep
It's quite common for kernel code to configure an event to resume
it, then suspend until the event fires. However, if the event fires
before the thread suspends itself, then it will essentially be
ignored and the thread will suspend for much longer than it planned.

This commit goes some way to preventing that by introducing a new
'drowsy' thread state. If a thread enters this state before configuring
the event, then any calls to resume by the event will move the thread
to a new 'insomniac' state. The next call to suspend while the thread
is insomniac will simply return (and set its state to runnable),
without suspending or sleeping.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>