memory: Moved remaining functionality into virtmem
This moves the last remaining functionality in the memory module
(bootstrapping the physical memory manager, initialising the virtual
memory manager and kernel heap, initialising the physical memory
manager, and giving the ability to modify the page mappings) into
the virtmem crate. It's where the code belonged all along, but it
needed to be moved carefully.

There is no meaningful functionality change.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>