virtmem: Replace kernel_pml4 with with_page_tables + refactor
This commit includes two changes. The first is replacing kernel_pml4
(which returned a mutable reference to the kernel's level 4 page
tables) with a new function with_page_tables, which takes a
closure acting on the page tables. The key difference here is
that with_page_tables synchronises using the mutex protecting the
kernel page table address. This means that with_page_tables ensures
there can never be more than one mutable reference (or one mutable
plus any immutable) to the page tables, which is a big safety gain.
It also gives us an opportunity to add extra safety checks in
future changes.

Secondly, the above change uncovered that device initialisation
did indeed result in multiple simultaneous mutable references, as
the new code resulted in a deadlock. This was due to a well-intended
but over-zealous attempt to use more dependency injection, rather
than accessing the kernel_pml4 directly. Now that things are much
more decoupled, we can sensibly undo that. Now, MMIO (which was
the culpret) uses with_page_tables directly, rather than needing
to be passed a mapper. This also means MMIO::Region::map doesn't
need to be unsafe any more, which is nice.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>