virtmem: Added map_pages abstraction over with_page_tables
Now that we've replaced kernel_pml4 with with_page_tables (see
a9ce74b), we can add another layer of abstraction for most use
cases, which is simply mapping a set of virtual memory pages
to arbitrary physical memory frames. Not only does this cut
down on code repetition, but it gives us an opportunity to check
the pages being mapped to enforce other requirements.

There are a few remaining cases that need to use with_page_tables
directly, but at least that can be the exception.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>