meta: Make it obvious that we only use 'Cargo.toml' for config
Since we switched to Bazel, we wouldn't want people to think that
the presence of a Cargo.toml file means the kernel can be built
with Cargo. This commit adds a disclaimer to that fact and renames
the file to ograC.toml to ensure it's not used accidentally by
tools.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>