tools: Added update-deps to automate updating dependencies
For now this only supports updating Rust and its tools to different
nightly versions. By default it updates to the first of the current
month. Alternatively the date can be specified with the -version
flag. There's no dry run option, as the repo is versioned, so we
can undo unwanted changes easily.

The default mode if a command is not specified is to run all update
commands. In future, more commands will be added to update other
parts of the repo.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>