bazel: Moved Rust dependencies into a macro
As the WORKSPACE file grows, it becomes increasingly difficult to
understand our Bazel dependencies. The long-term solution to this
is bzlmod, but it's hard to tell how long it will be before that
stabilises. Until then, we should use the deps.bzl pattern, moving
different sets of dependencies into separate macros. This commit
does that for our Rust dependencies (as the largest dependency set).
It also moves the checksums, crate overrides, and crate list into
variables, which will make them easier to modify over time.

Note that this does not update the update-deps tool. That will
happen in the next commit.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>