update-deps: Refactored and improved Starlark parsing
The previous code worked, but there was lots of repetition and it
was prone to subtle errors as things changed over time. This adds
helper functions and types to make parsing Starlark more predictable
and more concise. It also refactors the Starlark parsing into a
separate function for each file.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>