network: Wake initial workloads immediately if we have DHCP
The original design for the initial workloads functionality was
that one or more initial workloads would be registered (ideally,
before preparing any network interfaces). Then, when an interface
negotiates a DHCP configuration, the registered workloads would
be woken. This works fine, but it requires the initial workloads
to be set up before registering any network interfaces (unless
you're willing to wait for the configuration to expire and be
renegotiated).

This commit means we check whether we already have any DHCP
configurations already when registering the workload. If so, we
wake the workload and return immediately. This means we can
register initial workloads at any point and still get the
intuitive behaviour.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>