virtmem: Added map_frames_to_pages to remove mapping from MMIO
Previously, we used the map_pages helper function in all places
where we map pages, apart from when we initially remap the kernel
and when we map MMIO pages. The first exception is because we
need much more control than usual over the page mappings, which
is fine. The second exception is because the MMIO mappings need
to map to specific physical frames, rather than whatever the
allocator happens to have free.

This commit solves that by adding a second (almost identical)
helper function, map_frames_to_pages, which takes a frame iterator.
This means we can remove the custom mapping code from MMIO.

This may not seem like a big deal, but it means the virtmem crate
can now take ownership of mappings, allowing us to apply checks to
the pages that are being mapped.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>