virtmem: Check whether page mappings are frozen once per call
This is clearly a premature optimisation, but it's too obvious to
pass up. Rather than performing an atomic load on a variable that
is almost guaranteed to return true for each page we load, check
once at the top of map_(frames_to_)pages and skip the page check
if the mappings aren't frozen.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>