multitasking: Moved thread-switching assembly to crate root
Previously, the assembly code (and the Rust function declarations)
were part of the thread module, as most of the functions were used
in that module to start new threads. Only the switch_stack function
was used elsewhere, in the scheduler module. However, it was a bit
awkward to split them across modules like this.

This commit moves them into the top-level crate, effectively as a
sibling module.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>