physmem: Added an arena allocator, with allocation tracking
We already have a physical memory allocator, which works well.

This commit adds two new types; an allocation tracker and an
arena allocator. The tracker behaves similarly to an allocator,
but it simply keeps track of the allocations and deallocations
passed to it. This makes it possible to group together the
allocations made for a particular purpose. An allocation tracker
is created from an allocator to ensure they both track the same
areas of memory. If this were not the case, it would be possible
for the allocator to return a frame that the tracker was not
expecting to track.

Secondly, this adds the arena allocator. This is a short-lived
type that mutably borrows a generic frame allocator + deallocator,
plus an allocation tracker. The arena allocator is a frame
allocator + deallocator, which defers entirely to the generic
allocator it borrows. The key difference is that before returning
the allocated frame (or before deallocating the deallocated frame),
the frame is passed to the allocation tracker to update its records.
Finally, the arena allocator has a deallocate_all_frames method,
which can be called to deallocate all frames tracked by the tracker.

Combining the two types together will allow us to perform allocations
as normal, but then to deallocate all frames allocated for a particular
arena. The plan will be for each user process to contain an allocation
tracker so that when the process exits, all of its (unique) allocations
can be deallocated reliably.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>