virtmem: Track frozen kernel mappings more efficiently
Once the kernel is initialised, we freeze the level 4 kernel page
mappings to ensure that all level 4 page tables agree on all page
mappings in kernelspace. Once the mappings are frozen, we check
that each new mapping that would affect kernelspace is in an area
of memory that is already present in the level 4 page tables.

Previously, we did this by using the page table and indexing into
the table. This works fine, but it's potentially an expensive
lookup.

This commit adds a fixed size bitmap, which is used to track the
kernelspace mappings and check them more efficiently.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>