spin: Added a detonating spinlock to help identify deadlocks
Deadlocks can be difficult to identify, as they may cause no
symptoms to begin with, increasing the distance between cause and
effect. It can also be difficult to identify which mutex is being
deadlocked, particularly when the deadlock occurs in an interrupt
handler, as the symptoms could then appear in entirely unrelated
non-interrupt code.

This commit adds a new spinlock implementation, which has two key
features. Firstly, it records the file and line number where it is
locked. Secondly, if an attempt to lock a mutex fails after many
attempts, the kernel will panic, printing both the file and line
of the code attempting to lock the mutex and the file and line of
the code that last successfully locked the mutex. This identifies
both which lock is being deadlocked and the potential cause.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>