spin: Switch from the public spin crate to our own spinlock
The previous commit introduced our own spinlock, which detonates
if an attempt to lock the mutex takes too long. This is to help
identify both current and future deadlocks. The biggest difference
is switching from a `lock()` method to a `lock!()` macro. This
allows us to capture the file:line where the lock attempt occurs,
to aid debugging if a deadlock does happen.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>