process: Added initial support for user processes
This doesn't add a vast amount of new code but it's a big change.
The thread structure now also includes the process thread id, the
kernel process id of the process that owns it, and the process's
page table (or `None` for kernel threads). It also makes the
create_user_thread function available only within the multitasking
crate. Instead other code should call create_user_thread on a
process.

Processes contain a mapping of its thread ids to kernel thread
ids, a helper for identifying the next process thread id. It also
its kernel process id, its page table, and its allocation tracker.
A process is created by passing in a byte slice containing an ELF
binary, which is parsed and mapped into memory (using the process's
page tables). An initial user thread is also created, with the
entry point specified in the ELF binary. This initial thread is
not started by default.

The scheduler has also been updated, so that when switching thread,
we either switch to the new thread's page table, or the kernel's
for kernel threads.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>