update-deps: Fixed formatting bug
If we end up updating any Go modules, we split the logic based on
whether there's one or more than one module updated, so we can use
the right pluralisation. However, the check incorrectly checked
whether 0 modules had been updated, which is made impossible by
an identical check further up in the function. This fixes that.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>