update-deps: Added the ability to have ignored fields
This can be useful for deliberately ignoring a field we don't care
about, as we otherwise require all fields to be specified. Note
that for now, the field still needs to have the type StringField.
Ideally, the type wouldn't matter, but refactoring that would take
more time than I have right now.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>