qemu: Make the image handling simpler and more robust
We have to make a copy of the bootable image, as Qemu refuses to
boot an image file that is read-only and Bazel makes dependency
files read-only to prevent the dependant from modifying its
dependency. To reduce the chance that the copy and the input have
the same name, we call the copy _image.bin.

At the same time, the command line arg to specify the bootable
image will always be the same and depends on the (internal) name
of the copied image. As a result, we now include that argument in
the qemu rule, rather than requiring the caller to supply it.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>