bazel: Use the rustc managed by Bazel to get LLVM-tools
We want to ensure that any tools using LLVM-tools (currently
@bootloader's build script and //tools/bootimage) use the LLVM
tools managed by Bazel and rules_rust. Since the llvm_tools crate
determines the path to the LLVM tools by calling `rustc --print sysroot`,
if a host version of rustc is found in the $PATH before the rustc
managed by Bazel, then it will use the host LLVM tools (if present)
or will fail to find them if absent. Therefore, we need to ensure
that Bazel's rustc appears before any host version in the $PATH.

This commit attempts to do just that. Hopefully this will fix the
issue, but I'll need to check it on another machine where rustup
hasn't installed llvm-tools.

See #12

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>