bazel: Added a rule to check the formatting of Go code
This adds a new gofmt rule, which can be used to check that a set
of Go code is correctly formatted, according to `gofmt -s -l`. If
the code is formatted, then the 'build' will succeed with no effect.
If not, the 'build' will fail, printing a list of filenames that
are not formatted. This uses the gofmt binary provided by rules_go.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>