process: Moved ELF binary parsing and validation into a module
This is self-contained functionality, which will be much easier
to test and extend separately from the process memory management
that follows. At the same time, this abstracts away the ELF-specific
functionality into a new Binary/Segment type pair. Not only does
this abstract ELF into one module, but it also simplifies the
processing in Process::create_user_process.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>