deps: Added a tool to help analyse the Rust crate dependency graph
The Rust crates we use are an important part of our dependency graph.
However, they can be tricky to analyse, due to the way they are
fetched by rules_rust. To help in that process, this new command
can be used to parse data from Bazel and rules_rust, producing a
series of information and warnings.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>