crypto: Implemented SHA-256
Currently, we use the SHA-256 implementation from RustCrypto, in
the sha2 crate. This functions correctly, but the code and API
are quite complex and introduce several otherwise unnecessary
dependencies.

While implementing cryptographic algorithms is a high risk activity,
the reduced dependency management burden and code complexity is a
good reason.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>