crypto: Implement ChaCha20
As with the previous commit, this is intended to replace the
ChaCha20 implementation from RustCrypto. The chacha20 crate, like
the sha2 crate, has complex code and a strange API. It also brings
in several otherwise unnecessary dependencies.

As before, implementing cryptographic algorithms is a high risk
activity, but it seems like a sensible trade-off here.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>