crypto: Fixed block caching in chacha20
Previously, we would mark the entire block cache as consumed. If
the data being crypted has a smaller size than the remaining key
stream in the block cache, then this would result in key stream
bytes being discarded.

This commit fixes the issue by decrementing the bytes remaining
in the block cache by the number of bytes crypted, rather than
setting it to zero.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>