loader: Abstract away ELF-specific aspects, require binary name
This commit includes two changes. Firstly, the ELF-specific aspects
of the loader are now abstracted away. Rather than calling Binary::parse_elf,
the caller uses Binary::parse, which defers to the elf crate itself.
This will allow us to add support for other binary formats in the
future, without needing to affect code beyond the loader.

Secondly, the new Binary::parse function requires not only the
binary's file contents but also its name, in case we add support
for binary formats that are dependent on the file extension.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>