qemu: Do not reboot in response to a triple fault
In exceptionally rare circumstances (mainly when we manage to
unmap the kernel somehow), it becomes possible to triple fault
the CPU. It may even become possible to triple fault without
even having our CPU exception handler being invoked. In such
circumstances, it can be very tricky to work out what has gone
wrong. This is only made worse by the virtual CPU rebooting
like a physical machine would, as it throws away the debugging
state.

This adds an extra flag to Qemu, asking it not to reboot in
response to a CPU error, such as a triple fault.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>