memory: Introduce new virtual and physical memory types
This is the first step of migrating the majority of the
kernel from using the x86-64 specific types to using
cross-platform types with per-platform behaviours.

Note that this only introduces new code, the new code is
not yet used anywhere.

There is some functionality that will always be specific
to a subset of platforms, such as control registers and
mechanisms for changing address space. These will always
need to use platform-specific types, such as `x86_64::VirtAddr`.
However, for most of the kernel, more abstract representations
for virtual and physical addresses, ranges of addresses,
pages of virtual memory, frames of physical memory, and
ranges of both. These can be implemented once for all
platforms, having platform-specific code where necessary,
guarded with #[cfg] attributes. For example, the specific
range of valid virtual addresses will vary between platforms,
but that can be constrained to a single function, with the
remaining functionality remaining cross-platform.

My moving to cross-platform types for the bulk of the
kernel, we make any future efforts to support new instruction
set architectures much easier.

Secondly, there are some aspects of the types currently used
from the `x86_64` crate that can be challenging to use, or
have already been duplicated. The prime example of this is
that virtual memory pages and physical memory frames are both
generic types over their size. This means that it is not easy
to have a container of heterogenious page/frame sizes. The
benefit of this approach is that it allows compile-time checks
for certain situations, such as requiring that a page table
always use a 4-KiB physical frame. The alternative is to store
the size in the value, rather than its type. This is more
flexible, as it allows heterogenious collections, but it
prevents the compile-time checks previously mentioned.

This first step introduces eight primary types, which are
described in `//kernel/memory/lib.rs`. Their correspondence
to the types they will replace are as follows:

- `PhysAddr` replaces `x86_64::PhysAddr`
- `PhysAddrRange` is new
- `PhysFrameSize` is new
- `PhysFrame` replaces `x86_64::structures::paging::PhysFrame`
- `PhysFrameRange` replaces `x86_64::structures::paging::frame::{PhysFrameRange, PhysFrameRangeInclusive}`
- `VirtAddr` replaces `x86_64::VirtAddr`
- `VirtAddrRange` replaces `//kernel/memory/memlayout::VirtAddrRange`
- `VirtPageSize` replaces `x86_64::structures::paging::{PageSize, Size4KiB, Size2MiB, Size1GiB}`
- `VirtPage` replaces `x86_64::structures::paging::Page`
- `VirtPageRange` replaces `x86_64::structures::paging::page::{PageRange, PageRangeInclusive}`
- `PhysFrameAllocator` replaces `x86_64::structures::paging::FrameAllocator`
- `PhysFrameDeallocator` replaces `x86_64::structures::paging::FrameDeallocator`

There are several themes that run through these new types, besides
those that have been included from the types they replace:

- Types that act on or relate to virtual memory have the `Virt` prefix.
- Types that act on or relate to physical memory have the `Phys` prefix.
- Corresponding physical and virtual types have identical APIs.
- Where possible, const functions and inlining are used to make APIs as natural and flexible as possible.
- Generics is avoided completely.
- No static data is stored; only types, functions/methods, and constants.

Future commits will add more functionality to //kernel/memory.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>