memory: Add memory constants using the new types
See d11ba81 for the context behind this set of changes.

This adds identical memory layout constants to those in //kernel/memory/memlayout.
The only difference is that these use the new memory types. Note
also that the `phys_to_virt_addr` function is added to the memory
module, not the new constants child module.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>