memory: Introduced functionality to read and write page tables
See d11ba81 for the context behind this set of changes.

This adds new functionality for using page tables. Specifically,
it allows us to:

- Map memory
- Unmap memory
- Change mapping flags
- Translate virtual to physical addresses

This draws inspiration from `x86_64::structures::paging::{MappedPageTable, OffsetPageTable}`,
but makes some changes. The biggest change (as described in d11ba81)
is that `VirtPage` and `PhysFrame` store their size in the value,
not the type. This means that we need not have separate traits for
handling mappings of different sizes. Most of the page table API is
unsafe, to reflect the fact that the safety requirements of page
table mappings cannot be encoded purely in the Rust type system.

The page table handling code makes strong use of unit tests, where
the offset at which all physical memory is already mapped is zero,
leading to the page tables being accessed directly in virtual
memory. This allows us to safely simulate functioning page tables
in userspace.

Note that as the page tables are the bridge between virtual and
physical memory, the relating types have neither a `Virt` nor `Phys`
prefix.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>