memory: Remove redundant memlayout crate
See d11ba81 for the context behind this set of changes.

This is the final commit as part of this switch. Now that we have
switched entirely over to the new virtual and physical memory types,
we no longer use the constants in the old types that existed in the
memlayout crate. As a result, we can remove it entirely.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>