memory: Fixed handling of alignment in address types
Contrary to both the documentation and my understanding, it is
indeed possible for an alignment to change a valid address into
an invalid address. Rather than blindly accept this, we now panic
in response to an invalid address. For lower risk versions, we
also offer a `try_` variant of each, which returns a result,
giving an error if the resulting address is invalid.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>