loader: Add the ability to load kernel binaries
This means a binary can either be loaded as a userspace-only binary,
where the previous restrictions still apply, or as a kernel-only or
hybrid binary, where program segments and the entry point can exist
outside userspace.

This also removes the assumption that program segments should always
allow access from userspace. Instead, we only add the flag if the
segment is in the lower half of memory.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>