virtmem: Parse the kernel binary to identify mappings precisely
Previously, we essentially guessed the purpose of each page mapping
within the kernel binary, by taking addresses of various kinds of
value and comparing them to the mappings. This was imprecise enough
that we had to have a KernelBinaryUnknown type, which felt dirty
and unnecessary.

This takes a slightly more expensive but much more precise approach.
Instead, we parse the kernel binary from memory, using the program
segments to identify precisely where each mapping belonged, then
using the page table flags to reverse-engineer the intended purpose.
It means processing more data, but it's a much more robust approach.

This does mean we can't disambiguate between the kernel's constants
and its strings, so this removes KernelStrings and the now-redundant
KernelBinaryUnknown.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>