bootimage: Replaced Rust + LLVM-tools + GDB version with pure Go
This replaces the bootimage tool (which turns the bootloader binary
and kernel into a bootable disk image) with a simpler implementation
written in Go. The old version was written in Rust and would use
LLVM-objcopy to flatten the ELF binary, then make another copy that
was a multiple of the block size (512 bytes). Finally, at build time,
if the host has gdb installed, it would be used to check that the
bootloader's second stage was small enough to be copied using simple
BIOS routines.

This new version is somewhat simpler. The old version needed to make
a copy of the input so that it could be make writable as llvm-objcopy
duplicates the input flags to the output. Sicne we control the output
creation in the new version, we can set the output flags correctly
without needing to modify the input. Similarly, we can pad the output
to the block size without needing to copy it first. Finally, we can
check the symbol table to check the second stage bootloader while we
parse the ELF binary to extract the contents.

All in all, we need fewer dependencies (just the Go standard library),
and we can do everything in a single pass, rather than needing to copy
both the input and the output. It also more clearly encodes the
understanding of what's happening, making future changes and adaptations
easier.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>