bazel: Switch to pre-built standard library
Now that we have the pre-built copies of the Rust standard library
for x86_64-unknown-none, we can use those instead of fetching libcore
and liballoc with the host copy of Rust and fetching compiler_builtins
directly. We also don't need to build them any more.

This takes us part of the way to getting cross-compiling for free
with rules_rust.

See #14.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>