firefly: Added helper functions for calling raw syscalls
The syscalls module will call lots of different syscalls, and we
don't want to rewrite the assembly every time. Note that we make
heavy use of named registers for both inputs and outputs, relying
on the compiler to make this as efficient as possible. We also
document the two registers that get clobbered.

With the helpers defined, we can reimplement exit_thread by
calling syscall0.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>