kernel: Implement print_message and print_error syscalls
This adds two more syscalls, both for printing messages. The first
writes to the standard output, the second to the standard error.
For now, both print to the kernel's output. To differentiate them
from the kernel's output, both use ANSI terminal colour codes;
print_message prints in green, print_error in red. If the message
is not entirely valid UTF-8, we return the new IllegalParameter
error.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>