plan: Added the abstract syntax tree types for the Plan language
Plan is a minimal Lisp designed to be used to specify the ABI of
an operating system. With the ABI defined in a Plan document,
compatible code can be generated to either call or answer the
corresponding system calls. This ensures that user and kernel
code do not have different interpretations of the same data in
memory.

This adds types to represent the various nodes and expressions in
a Plan document.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>