plan: Added an interpreter for the Plan interface language
Plan is a minimal Lisp designed to be used to specify the ABI of
an operating system. With the ABI defined in a Plan document,
compatible code can be generated to either call or answer the
corresponding system calls. This ensures that user and kernel
code do not have different interpretations of the same data in
memory.

This adds an interpreter, which type-checks the abstract syntax
tree and returns a validated high-level representation of a Plan
document. This includes ensuring that all structures are aligned
in memory for the chosen instruction set architecture and that
all referenced types are defined.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>