plan: Added an auto-formatter for the Plan interface language
Plan is a minimal Lisp designed to be used to specify the ABI of
an operating system. With the ABI defined in a Plan document,
compatible code can be generated to either call or answer the
corresponding system calls. This ensures that user and kernel
code do not have different interpretations of the same data in
memory.

This includes an auto-formatter, which takes an abstract syntax
tree and produces a Plan document containing the same information
in the standard format. This also includes functionality to sort
the fields defined in a Plan document into the standard order.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>