plan: Added the plan tool, for operating on the Plan language
Plan is a minimal Lisp designed to be used to specify the ABI of
an operating system. With the ABI defined in a Plan document,
compatible code can be generated to either call or answer the
corresponding system calls. This ensures that user and kernel
code do not have different interpretations of the same data in
memory.

This implements the plan CLI, which can be used to validate or
auto-format a Plan document, and to generate user or kernel code
in Rust from a Plan document. For now, the user code is complete,
in that it produces definitions for all enumerations and structures,
including the synthetic enumeration for the set of system calls,
plus a function definition for each system call, which can be used
directly. However, the kernel code only defines the enumeration and
structure definitions and the syntetic enumeration for the set of
system calls. The system calls themselves are not defined. This
will be added once I work out the most ergonomic way to type-check
the system call implementations without introducing performance
issues.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>